Fix data race issue showed in the unit tests
Concurrency call used in the unit test with share variables was causing
the data issue in the unit test. Proper lock/unlock added to solve the
issue. Similar issue verify in the cache.go causing race condition in the
main.go. lock/unlock place in the WriteMetric in the autoupdating.go